backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
httpx==0.27.0
pytest==7.4.4
pytest-asyncio==0.21.1
openpyxl==3.1.2
python-dotenv==1.0.1
PyMuPDF==1.24.9
pillow==10.4.0
azure-ai-formrecognizer==3.3.1
azure-core==1.30.1
